pyproject.toml
[tool.poetry]
name = "atomic-habits-drf"
version = "0.1.0"
description = ""
authors = ["Krimsi13"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
black = "^24.4.2"
setuptools = "^70.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
